+++ pkg/metrics/metrics.go
@@ -36,23 +36,25 @@ type Server struct {
 
 // Start starts serving the metrics server.
 func (s *Server) Start() {
-	log.Info(nil, "Metrics server started.", zap.String("address", s.Addr))
+	log.Info(nil, "Metrics server started", zap.String("address", s.Addr))
 
-	if err := s.ListenAndServe(); err != http.ErrServerClosed {
-		log.Error(nil, "Metrics server died unexpected.", zap.Error(err))
+	if err := s.ListenAndServe(); err != nil {
+		if err != http.ErrServerClosed {
+			log.Error(nil, "Metrics server died unexpected", zap.Error(err))
+		}
 	}
 }
 
 // Stop terminates the metrics server gracefully.
 func (s *Server) Stop() {
-	log.Debug(nil, "Start shutdown of the metrics server.")
+	log.Debug(nil, "Start shutdown of the metrics server")
 
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
 	defer cancel()
 
 	err := s.Shutdown(ctx)
 	if err != nil {
-		log.Error(nil, "Graceful shutdown of the metrics server failed.", zap.Error(err))
+		log.Error(nil, "Graceful shutdown of the metrics server failed", zap.Error(err))
 	}
 }